chenged the sizes of the reserved fields within TRDP_ETB_CTRL_VDP_T to comply with 32 and 16 bit alignment
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@1257 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ src/api/tau_tti_types.h
@@ -197,8 +197,8 @@ typedef struct
 
 typedef struct
 {
-    UINT16                  reserved01;     /**< reserved (=0) */
-    UINT32                  reserved02;     /**< reserved (=0) */
+    UINT32                  reserved01;     /**< reserved (=0) */
+    UINT16                  reserved02;     /**< reserved (=0) */
     TRDP_SHORT_VERSION_T    userDataVersion;/**< version of the vital ETBCTRL telegram
                                              mainVersion = 1, subVersion = 0 */
     UINT32                  safeSeqCount;   /**< safe sequence counter, as defined in B.9 */
